avoid 'array subscript has type ‘char’' warning

+++ json.c
@@ -476,7 +476,7 @@ json_value * json_parse_ex (json_settings * settings, const json_char * json, ch
 
                      default:
 
-                        if (isdigit (b) || b == '-')
+                        if (isdigit ((int)b) || b == '-')
                         {
                            if (!new_value (&state, &top, &root, &alloc, json_integer))
                               goto e_alloc_failure;
@@ -550,7 +550,7 @@ json_value * json_parse_ex (json_settings * settings, const json_char * json, ch
             case json_integer:
             case json_double:
 
-               if (isdigit (b))
+               if (isdigit ((int)b))
                   continue;
 
                if (b == 'e' || b == 'E')
